+++ chart/compass/charts/external-services-mock/templates/oauth-credentials-secret.yaml
@@ -52,7 +52,7 @@ type: Opaque
 data:
   {{- $key := .Values.global.director.selfRegister.secret.key }}
   {{- with .Values.global.director.selfRegister.local.templateMappings }}
-  {{ $key }}: {{ printf `{"local-region":{%s,%s,%s,%s,%s,%s}}` (tpl .clientIDMapping $) (tpl .clientSecretMapping $) (tpl .urlMapping $) (tpl .tokenURLMapping $) (tpl .x509CertificateMapping $) (tpl .x509KeyMapping $) | b64enc | quote }}
+  {{ $key }}: {{ printf `{"eu-1":{%s,%s,%s,%s,%s,%s}}` (tpl .clientIDMapping $) (tpl .clientSecretMapping $) (tpl .urlMapping $) (tpl .tokenURLMapping $) (tpl .x509CertificateMapping $) (tpl .x509KeyMapping $) | b64enc | quote }}
   {{- end }}
 ---
 apiVersion: v1
